Add common module and move upload_stream and wrap_stream there
